+++ ydb/core/tx/datashard/build_index/secondary_index.cpp
@@ -18,9 +18,10 @@
 #include <yql/essentials/public/issue/yql_issue_message.h>
 
 #include <ydb/core/ydb_convert/ydb_convert.h>
+#include <ydb/core/ydb_convert/table_description.h>
+
 #include <util/generic/algorithm.h>
 #include <util/string/builder.h>
-#include <ydb/core/ydb_convert/table_description.h>
 
 namespace NKikimr::NDataShard {
 
@@ -630,271 +631,6 @@ void TDataShard::HandleSafe(TEvDataShard::TEvBuildIndexCreateRequest::TPtr& ev, 
         StartScan(this, std::move(scan), id, seqNo, rowVersion, userTable.LocalTid);
     } catch (const std::exception& exc) {
         FailScan<TEvDataShard::TEvBuildIndexProgressResponse>(id, TabletID(), ev->Sender, seqNo, exc, "TBuildIndexScan");
-    }
-}
-
-class TValidateUniqueIndexScan: public NTable::IScan {
-public:
-    TValidateUniqueIndexScan(
-        ui64 buildIndexId,
-        const TUserTable& tableInfo,
-        TProtoColumnsCRef targetIndexColumns,
-        const TScanRecord::TSeqNo& seqNo,
-        ui64 dataShardId,
-        NActors::TActorSystem* actorSystem,
-        NActors::TActorId responseActorId)
-        : BuildIndexId(buildIndexId)
-        , SeqNo(seqNo)
-        , DataShardId(dataShardId)
-        , ActorSystem(actorSystem)
-        , ResponseActorId(responseActorId)
-        , ScanTags(BuildTags(tableInfo, targetIndexColumns))
-        , IndexColumnNames(targetIndexColumns.begin(), targetIndexColumns.end())
-    {
-    }
-
-    TInitialState Prepare(IDriver*, TIntrusiveConstPtr<TScheme> scheme) override {
-        Scheme = std::move(scheme);
-        MakeTypeInfos();
-        return {EScan::Feed, {}};
-    }
-
-    EScan Seek(TLead& lead, ui64) override {
-        lead.To(ScanTags, {}, NTable::ESeek::Lower);
-        return EScan::Feed;
-    }
-
-    TString DuplicateKeyErrorMessage() const {
-        TStringBuilder res;
-        res << "Duplicate key found: (";
-        for (size_t i = 0; i < IndexColumnNames.size(); ++i) {
-            if (i > 0) {
-                res << ", ";
-            }
-            res << IndexColumnNames[i] << "=";
-            DbgPrintValue(res, LastIndexKey->GetCells()[i], IndexColumnTypeInfos[i].ToTypeInfo());
-        }
-        res << ")";
-        return std::move(res);
-    }
-
-    EScan Feed(TArrayRef<const TCell> key, const TRow& row) override {
-        if (row.Size() != ScanTags.size()) {
-            return FinishValidation(NKikimrIndexBuilder::EBuildStatus::ABORTED, TStringBuilder() << "Row size mismatch: expected " << ScanTags.size() << ", got " << row.Size());
-        }
-
-        ++ReadRows;
-        ReadBytes += CountRowCellBytes(key, *row);
-
-        TArrayRef<const TCell> rowCells = *row;
-        if (!FirstIndexKey) {
-            FirstIndexKey.emplace(rowCells);
-            LastIndexKey.emplace(rowCells);
-            return EScan::Feed;
-        }
-
-        if (TypedCellVectorsEqualWithNullSemantics(rowCells.data(), LastIndexKey->GetCells().data(), IndexColumnTypeInfos.data(), IndexColumnTypeInfos.size()) == ETriBool::True) {
-            return FinishValidation(NKikimrIndexBuilder::EBuildStatus::BUILD_ERROR, DuplicateKeyErrorMessage());
-        }
-        LastIndexKey.emplace(rowCells);
-        return EScan::Feed;
-    }
-
-    TAutoPtr<IDestructable> Finish(const std::exception& ex) override {
-        Issues.AddIssue(TStringBuilder() << "Scan failed: " << ex.what());
-        return Finish(EStatus::Exception);
-    }
-
-    TAutoPtr<IDestructable> Finish(EStatus status) override {
-        if (StatusCode == NKikimrIndexBuilder::EBuildStatus::INVALID) {
-            if (status == EStatus::Exception) {
-                StatusCode = NKikimrIndexBuilder::EBuildStatus::BUILD_ERROR;
-            } else if (status != EStatus::Done) {
-                StatusCode = NKikimrIndexBuilder::EBuildStatus::ABORTED;
-            } else {
-                StatusCode = NKikimrIndexBuilder::EBuildStatus::DONE;
-            }
-        }
-
-        auto response = std::make_unique<TEvDataShard::TEvValidateUniqueIndexResponse>();
-        FillScanResponseCommonFields(*response, BuildIndexId, DataShardId, SeqNo);
-        auto& rec = response->Record;
-        rec.SetStatus(StatusCode);
-        rec.MutableMeteringStats()->SetReadRows(ReadRows);
-        rec.MutableMeteringStats()->SetReadBytes(ReadBytes);
-        if (Issues) {
-            NYql::IssuesToMessage(Issues, rec.MutableIssues());
-        }
-        if (FirstIndexKey) {
-            rec.SetFirstIndexKey(FirstIndexKey->GetBuffer());
-        }
-        if (LastIndexKey) {
-            rec.SetLastIndexKey(LastIndexKey->GetBuffer());
-        }
-        ActorSystem->Send(ResponseActorId, response.release());
-        return this;
-    }
-
-    EScan Exhausted() override {
-        return EScan::Final;
-    }
-
-    void Describe(IOutputStream& out) const override {
-        out << "TValidateUniqueIndexScan Id: " << BuildIndexId
-            << " Status: " << StatusCode << " Issues: " << Issues.ToOneLineString();
-    }
-
-    TString Debug() const {
-        TStringBuilder res;
-        Describe(res.Out);
-        return std::move(res);
-    }
-
-    void MakeTypeInfos() {
-        IndexColumnTypeInfos.reserve(ScanTags.size());
-        for (NTable::TTag tag : ScanTags) {
-            const NTable::TColInfo* colInfo = Scheme->ColInfo(tag);
-            Y_ENSURE(colInfo, "Column info not found for tag " << tag);
-            IndexColumnTypeInfos.emplace_back(colInfo->TypeInfo);
-        }
-    }
-
-    EScan FinishValidation(NKikimrIndexBuilder::EBuildStatus statusCode = NKikimrIndexBuilder::EBuildStatus::DONE, TString error = {}) {
-        StatusCode = statusCode;
-        Issues.AddIssue(error);
-        return EScan::Final;
-    }
-
-private:
-    const ui64 BuildIndexId;
-    const TScanRecord::TSeqNo SeqNo;
-    const ui64 DataShardId;
-    NActors::TActorSystem* const ActorSystem = nullptr;
-    const NActors::TActorId ResponseActorId;
-
-    TTags ScanTags;
-    std::vector<TString> IndexColumnNames;
-    TIntrusiveConstPtr<TScheme> Scheme;
-    std::vector<NScheme::TTypeInfoOrder> IndexColumnTypeInfos;
-
-    // Results
-    NKikimrIndexBuilder::EBuildStatus StatusCode = NKikimrIndexBuilder::EBuildStatus::INVALID;
-    NYql::TIssues Issues;
-    std::optional<TSerializedCellVec> FirstIndexKey;
-    std::optional<TSerializedCellVec> LastIndexKey;
-    ui64 ReadRows = 0;
-    ui64 ReadBytes = 0;
-};
-
-TAutoPtr<NTable::IScan> CreateValidateUniqueIndexScan(
-    ui64 buildIndexId,
-    const TUserTable& tableInfo,
-    TProtoColumnsCRef targetIndexColumns,
-    const TScanRecord::TSeqNo& seqNo,
-    ui64 dataShardId,
-    NActors::TActorSystem* actorSystem,
-    NActors::TActorId senderActorId)
-{
-    return new TValidateUniqueIndexScan(buildIndexId, tableInfo, targetIndexColumns, seqNo, dataShardId, actorSystem, senderActorId);
-}
-
-class TDataShard::TTxHandleSafeValidateUniqueIndexScan: public NTabletFlatExecutor::TTransactionBase<TDataShard> {
-public:
-    TTxHandleSafeValidateUniqueIndexScan(TDataShard* self, TEvDataShard::TEvValidateUniqueIndexRequest::TPtr&& ev)
-        : TTransactionBase(self)
-        , Ev(std::move(ev)) {
-    }
-
-    bool Execute(TTransactionContext&, const TActorContext& ctx) {
-        Self->HandleSafe(Ev, ctx);
-        return true;
-    }
-
-    void Complete(const TActorContext&) {
-        // nothing
-    }
-
-private:
-    TEvDataShard::TEvValidateUniqueIndexRequest::TPtr Ev;
-};
-
-void TDataShard::Handle(TEvDataShard::TEvValidateUniqueIndexRequest::TPtr& ev, const TActorContext&) {
-    Execute(new TTxHandleSafeValidateUniqueIndexScan(this, std::move(ev)));
-}
-
-void TDataShard::HandleSafe(TEvDataShard::TEvValidateUniqueIndexRequest::TPtr& ev, const TActorContext& ctx) {
-    auto& request = ev->Get()->Record;
-
-    const ui64 id = request.GetId();
-    TScanRecord::TSeqNo seqNo = {request.GetSeqNoGeneration(), request.GetSeqNoRound()};
-
-    try {
-        auto response = MakeHolder<TEvDataShard::TEvValidateUniqueIndexResponse>();
-        FillScanResponseCommonFields(*response, request.GetId(), TabletID(), seqNo);
-
-        LOG_N("Starting TValidateUniqueIndexScan TabletId: " << TabletID()
-            << " " << request.ShortDebugString());
-
-        auto badRequest = [&](const TString& error) {
-            response->Record.SetStatus(NKikimrIndexBuilder::EBuildStatus::BAD_REQUEST);
-            auto issue = response->Record.AddIssues();
-            issue->set_severity(NYql::TSeverityIds::S_ERROR);
-            issue->set_message(error);
-        };
-        auto trySendBadRequest = [&] {
-            if (response->Record.GetStatus() == NKikimrIndexBuilder::EBuildStatus::BAD_REQUEST) {
-                LOG_E("Rejecting TValidateUniqueIndexScan bad request TabletId: " << TabletID()
-                    << " " << request.ShortDebugString()
-                    << " with response " << response->Record.ShortDebugString());
-                ctx.Send(ev->Sender, std::move(response));
-                return true;
-            } else {
-                return false;
-            }
-        };
-
-        const auto tableId = TTableId(request.GetOwnerId(), request.GetPathId());
-        if (request.GetTabletId() != TabletID()) {
-            badRequest(TStringBuilder() << "Wrong shard " << request.GetTabletId() << " this is " << TabletID());
-        }
-        if (!IsStateActive()) {
-            badRequest(TStringBuilder() << "Shard " << TabletID() << " is " << State << " and not ready for requests");
-        }
-        if (!GetUserTables().contains(tableId.PathId.LocalPathId)) {
-            badRequest(TStringBuilder() << "Unknown table id: " << tableId.PathId.LocalPathId);
-        }
-        if (request.GetIndexColumns().empty()) {
-            badRequest("Empty index columns list");
-        }
-
-        if (trySendBadRequest()) {
-            return;
-        }
-
-        const auto& userTable = *GetUserTables().at(tableId.PathId.LocalPathId);
-        auto tags = GetAllTags(userTable);
-        for (auto column : request.GetIndexColumns()) {
-            if (!tags.contains(column)) {
-                badRequest(TStringBuilder() << "Unknown index column: " << column);
-            }
-        }
-
-        if (trySendBadRequest()) {
-            return;
-        }
-
-        TAutoPtr<NTable::IScan> scan = CreateValidateUniqueIndexScan(
-            id,
-            userTable,
-            request.GetIndexColumns(),
-            seqNo,
-            request.GetTabletId(),
-            TActivationContext::ActorSystem(),
-            ev->Sender);
-
-        StartScan(this, std::move(scan), id, seqNo, std::nullopt, userTable.LocalTid);
-    } catch (const std::exception& exc) {
-        FailScan<TEvDataShard::TEvValidateUniqueIndexResponse>(id, TabletID(), ev->Sender, seqNo, exc, "TValidateUniqueIndexScan");
     }
 }
